Alteração do arquivo que realiza as operações matemáticas, agora com números folgados

+++ operacoes.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 
 int main(){
-    int n1,n2,so,su,mt,dv;
+    float n1,n2,so,su,mt,dv;
 
     printf("Digite um número e tecle enter\n");
-    scanf("%d",&n1);
+    scanf("%f",&n1);
 
     printf("Digite outro número e tecle enter\n");
-    scanf("%d",&n2);
+    scanf("%f",&n2);
 
     so = n1 + n2;
     su = n1 - n2;
@@ -16,10 +16,10 @@ int main(){
 
     printf("\n--------------------------------------\n");
     printf("\n----------Veja os resultados----------\n");
-    printf("\n Soma: %d                             \n",so);
-    printf("\n Subtração: %d                        \n",su);
-    printf("\n Multiplicação: %d                    \n",mt);
-    printf("\n Divisão: %d                          \n",dv);
+    printf("\n Soma: %2.2f                             \n",so);
+    printf("\n Subtração: %2.2f                        \n",su);
+    printf("\n Multiplicação: %2.2f                    \n",mt);
+    printf("\n Divisão: %2.2f                          \n",dv);
     printf("\n--------------------------------------\n");
 
     return 0;
